requirements.txt
airflow-dbt
dbt-snowflake
boto3
pandas
great_expectations 
airflow-provider-great-expectations
retrying
sqlalchemy
snowflake-sqlalchemy
apache-airflow-providers-snowflake
snowflake-connector-python
apache-airflow[amazon]

